pcp: Fix bug when muting layers/removing sublayers with sublayers
When computing incremental changes, Pcp needs to ignore the layers
being muted/removed *and* any of the sublayers they bring in when
recomputing state because these layers should no longer contribute
opinions to the composed scene. The latter was not happening, which
caused prim indexes to not be invalidated as expected. This fix just
ensures we do track all of those layers.

Fixing this issue revealed a separate bug when invalidating indexes
for prims that reference a muted layer. The logic for determining
when this invalidation was necessary was flipped and also did not
account for culled dependencies. This has been fixed and new unit
tests added to cover this case.

(Internal change: 2345458)